govc: Add json support
`export GOVC_VI_JSON=TRUE` to switch XML communication to JSON.

This is experimental attempt to get json working in govc. It requires a
vCenter 8.0.1.0 after the Beta release.

There are not bat tests yet as vcsim does not yet know how to talk json.

Testing done: Run `datacenter.info` and `folder.info` against a vCenter
server. Observed the network traffic with `-trace` - it is JSON and the
communication works.