fix: support `binary` and `dateTime` in JSON
This change adds support for `binary` and `dateTime` in JSON.

For `binary` golang `[]byte` JSON serialization matches the vCenter
format. Thus adding the name ot the maps was enough.

For `dateTime` the format of `time.Time` `Marshaler` and `Unmarshaler`
matches vCenter format. The discriminator code required support for
custom marshaling logic. In addition the name had to be mapped.

Closes: #3127