Support array of enum in JSON
govmomi does not emit ArrayOf<enum> types. However this type is
possible in the type system.

This change adds support for any types to be used with "ArrayOf" prefix.
It uses go slice types to represent "ArrayOf*" types. On deserialization
if matching type name in the map is not found attempt is made with
"ArrayOf" a prefix. On serialization if the type of any is slice or
array we will form the name from "ArrayOf" prefix and the actual type
name.